requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
accelerate==1.13.0
    # via -r requirements.in
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   tqdm
    #   uvicorn
diffusers==0.37.1
    # via -r requirements.in
dnspython==2.8.0
    # via
    #   email-validator
    #   pymongo
dotenv==0.9.9
    # via -r requirements.in
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi[standard]==0.135.3
    # via -r requirements.in
fastapi-cli[standard]==0.0.24
    # via fastapi
fastapi-cloud-cli==0.16.1
    # via fastapi-cli
fastar==0.10.0
    # via fastapi-cloud-cli
filelock==3.25.2
    # via
    #   diffusers
    #   huggingface-hub
    #   torch
fsspec==2026.3.0
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.4.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   diffusers
    #   fastapi
    #   fastapi-cloud-cli
    #   huggingface-hub
huggingface-hub==1.10.1
    # via
    #   accelerate
    #   diffusers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==9.0.0
    # via diffusers
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
loguru==0.7.3
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.4.4
    # via
    #   accelerate
    #   diffusers
    #   transformers
packaging==26.0
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
pillow==12.2.0
    # via diffusers
psutil==7.2.2
    # via accelerate
pydantic[email]==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.11.2
    # via fastapi
pydantic-settings==2.13.1
    # via fastapi
pygments==2.20.0
    # via rich
pymongo==4.16.0
    # via -r requirements.in
python-dotenv==1.2.2
    # via
    #   dotenv
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.26
    # via fastapi
pyyaml==6.0.3
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2026.4.4
    # via
    #   diffusers
    #   transformers
requests==2.33.1
    # via diffusers
rich==15.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.19.7
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
safetensors==0.7.0
    # via
    #   accelerate
    #   diffusers
    #   transformers
sentry-sdk==2.57.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
starlette==1.0.0
    # via fastapi
sympy==1.14.0
    # via torch
tokenizers==0.22.2
    # via transformers
torch==2.11.0
    # via accelerate
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   transformers
transformers==5.5.3
    # via -r requirements.in
typer==0.24.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
    #   huggingface-hub
    #   transformers
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   starlette
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.44.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
win32-setctime==1.2.0
    # via loguru
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
